Move set & map to modules.
