+++ libstainless/Cargo.toml
@@ -8,7 +8,6 @@ categories = ["development-tools"]
 
 [lib]
 name = "stainless"
-path = "lib.rs"
 test = false
 doctest = false
